Also rearrange coreSet contents
