+++ package.json
@@ -112,8 +112,8 @@
     "react-transform-hmr": "^1.0.4",
     "redbox-react": "^1.5.0",
     "webpack-dev-middleware": "^1.12.0",
-    "webpack-hot-middleware": "^2.19.1",
-    "webpack-dev-server": "^2.9.1"
+    "webpack-dev-server": "^2.9.1",
+    "webpack-hot-middleware": "^2.19.1"
   },
   "optionalDependencies": {
     "canvas": "^1.6.7"
